requirements.txt
pandas>=2.0.0
gspread>=5.10.0
google-auth>=2.10.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
pyyaml>=6.0
google-cloud-bigquery>=3.12.0
pyarrow>=12.0.0
pandas-gbq>=0.26.1
uuid>=1.0.0
pytz>=2023.3